moved notifications from modal to settings page
