+++ packages/app/components/Settings/ProfileSettings.tsx
@@ -1,0 +1,5 @@
+import { ReactElement } from "react";
+
+export default function ProfileSettings(): ReactElement {
+  return <div>ligma</div>;
+}
