+++ packages/app/components/common/AvatarWithInitials.tsx
@@ -6,12 +6,14 @@ import getInitialsFromName from "../../utils/NameUtils";
 type AvatarWithInitalsProps = {
   name: string;
   fontSize?: number;
+  style?: any;
   [x: string]: any;
 };
 
 export default function AvatarWithInitals({
   name,
   fontSize,
+  style,
   ...props
 }: AvatarWithInitalsProps): ReactElement {
   return (
@@ -19,6 +21,7 @@ export default function AvatarWithInitals({
       style={{
         backgroundColor: name ? nameToRGB(name) : "#1abc9c",
         fontSize,
+        ...style,
       }}
       {...props}
     >
